#19 - reorganize files
